Fix timer class to track time if app in background
Timer class stops receiving updates when another app is opened or the
phone is locked. Thus, my stopwatch class would stop receiving
updates to increment the time when the user no longer had the app open.
Therefore my stopwatch class no longer calculates time based on the
timer class. It now calculates time based on the timestamp when the
timer first started and the timestamp of the current time. If the timer
is paused, it calculates the amount of elapsed time to subtract from
the time calculation.